requirements.txt
backoff==1.11.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
Deprecated==1.2.13
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
googleapis-common-protos==1.56.2
greenlet==1.1.2
grpcio==1.46.3
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.1
opentelemetry-api==1.11.1
opentelemetry-exporter-jaeger==1.11.1
opentelemetry-exporter-jaeger-proto-grpc==1.11.1
opentelemetry-exporter-jaeger-thrift==1.11.1
opentelemetry-exporter-otlp==1.11.1
opentelemetry-exporter-otlp-proto-grpc==1.11.1
opentelemetry-exporter-otlp-proto-http==1.11.1
opentelemetry-instrumentation==0.31b0
opentelemetry-instrumentation-flask==0.31b0
opentelemetry-instrumentation-kafka-python==0.31b0
opentelemetry-instrumentation-requests==0.31b0
opentelemetry-instrumentation-wsgi==0.31b0
opentelemetry-proto==1.11.1
opentelemetry-sdk==1.11.1
opentelemetry-semantic-conventions==0.31b0
opentelemetry-util-http==0.31b0
protobuf==3.20.1
psycopg2-binary==2.9.3
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.37
thrift==0.16.0
typing-extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
